requirements.txt
certifi==2019.3.9
chardet==3.0.4
docker==3.7.2
docker-pycreds==0.4.0
idna==2.8
psutil==5.6.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
rfc5424-logging-handler==1.4.1
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
websocket-client==0.56.0
